moved squad.py and format changes
